# packages/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Equipment Maintenance Hub API
  description: >
    Internal maintenance operations API for tracking equipment work orders
    through their lifecycle, from initial report through completion.
  version: 0.1.0
paths:
...
  /api/auth/login:
    post:
      operationId: login
      summary: Log in with email and password
      description: Exchanges valid credentials for a bearer token and the authenticated user.
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Credentials were valid; a bearer token was issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: The email or password is incorrect.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/auth/me:
    get:
      operationId: getCurrentUser
      summary: Get the authenticated user
      tags:
        - auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The user identified by the bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: The bearer token is missing, invalid or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token issued by POST /api/auth/login, sent in the Authorization header as "Bearer <token>".
  schemas:
...
    Role:
      type: string
      description: Access level of an authenticated user.
      enum:
        - technician
        - supervisor
        - admin
    User:
      type: object
      description: An authenticated user of the API. Never includes credentials.
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the user.
        email:
          type: string
          format: email
          description: Login email address of the user.
        name:
          type: string
          description: Full name of the user.
        role:
          $ref: '#/components/schemas/Role'
        technicianId:
          type: string
          format: uuid
          nullable: true
          description: Identifier of the technician profile linked to this user, if any.
      required:
        - id
        - email
        - name
        - role
        - technicianId
    LoginRequest:
      type: object
      description: Credentials used to obtain a bearer token.
      properties:
        email:
          type: string
          format: email
          description: Login email address.
        password:
          type: string
          description: Plain-text password.
      required:
        - email
        - password
    LoginResponse:
      type: object
      description: Result of a successful login.
      properties:
        token:
          type: string
          description: Bearer token to send in the Authorization header.
        user:
          $ref: '#/components/schemas/User'
      required:
        - token
        - user
    ApiError:
      type: object
      description: Standard error response shape.
      properties:
        message:
          type: string
          description: Human-readable error message.
        details:
          type: object
          description: Optional additional error context.
          additionalProperties: true
      required:
        - message